rename Hero
